Moved use_none to its own file, added test
